pyproject.toml
[project]
name = "day-3-lab-chatbot-vs-react-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.70.0",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.2",
]
